Repository URL checked before registering
Powershell does not allow duplicate PS repositories with the same URL.
This causes problems when the repo URL is already registered under a
different name. The code previously checked if the PS Repoistory was
registered by only looking at the name, not the URL.

This commit now checks the URL instead of the name. If the URL is not
found, it registers the URL with a guid as the name and then removes it
at the end of the script.